student_Admission_portal/student_admissions/app/requirements.txt
Flask>=2.1
Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
WTForms
Flask-Login
Flask-Mail
python-dotenv
twilio
email-validator
gunicorn
psycopg2-binary
Flask-Cors
requests
